TOOLS-1826 TOOLS-2403 Fix demux deadlock and panic
Prior to this commit, the archive demultiplexer receivers implemented
two interfaces, DemuxOut and intent.file, both of which have a Close
method.  While Close was protected with a sync.Once, this was not enough
to prevent problem behavior.

When the RegularCollectionReceiver Write method blocks waiting for a
reader, calling Close (e.g. from the intent side after an error) would
panic.  If the intent side Close is called after all documents in a
block are read, but before EOF is read, a deadlock would occur.

This commit changes the DemuxOut interface to replace the Close method
with a new End method, so that the demux side and the intent side have
distinct terminations and only these are used to signal to the other
side.

NOTE: This is an API breaking change.  Much of it is internal to the
operation of the demultiplexer, but this change does mean that intents
must be closed on error or EOF or the demultiplexer will not terminate.